% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_lms.R
\name{bp_lm}
\alias{bp_lm}
\title{Pie regression}
\usage{
bp_lm(x, left = TRUE, break_dates, tvlm = FALSE, ...)
}
\arguments{
\item{x}{\code{lm} object. It is the global regression model.}

\item{left}{\code{logical}. By default set to \code{TRUE}, i.e. the breakdate is the end date of each submodel}

\item{break_dates}{optional, to indicate the breakup dates if they are known.}

\item{tvlm}{By default set to \code{FALSE}. Indicates which model will be run on each sub data. \code{FALSE} means a \link{lm} will be run.}

\item{...}{other arguments passed to \code{\link[tvReg:tvLM]{tvReg::tvLM()}}.}
}
\value{
Returns an element of class \code{bp_lm}. It is a list containing the following elements:
\item{model}{all computed models, each of class \code{lm} or \code{tvlm} according to the parameter specified above}
\item{start}{start date of the time serie}
\item{end}{end date of the time serie}
\item{frequency}{frequency of the time serie}
\item{breakdates}{a list of the breakup dates}
\item{left}{same as the parameter specified above}
\item{tvlm}{same as the parameter specified above}
}
\description{
Computes as many regressions as breakup dates within a global model
}
