% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_reg.R
\name{break_data}
\alias{break_data}
\title{Break data}
\usage{
break_data(x, break_dates, right = TRUE, ...)
}
\arguments{
\item{x}{a \code{mts} to split}

\item{break_dates}{the date(s) at which you want to divide the data}

\item{right}{\code{logical}. By default set to \code{TRUE}, i.e. the breakdate is the end date of each subcolumn}
}
\value{
a \code{mts} containing as many times more data columns than breakdates
}
\description{
Splits a database according to one (or more) date
}
