% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_hansen.R
\name{deprecated-tvCoef}
\alias{deprecated-tvCoef}
\alias{hansen.test}
\alias{bp.lms}
\title{Deprecated function}
\usage{
hansen.test(...)

bp.lms(...)
}
\arguments{
\item{...}{parameters}
}
\description{
Deprecated function
}
