% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{full_exogeneous_matrix}
\alias{full_exogeneous_matrix}
\title{Extract Full Transformed Exogoneous matrix}
\usage{
full_exogeneous_matrix(model, ...)
}
\arguments{
\item{model}{the model}

\item{...}{other unused parameters}
}
\description{
Extract Full Transformed Exogoneous matrix
}
