% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\alias{get_data.lm}
\alias{get_data.dynlm}
\alias{get_data.tvlm}
\alias{get_data.bp_lm}
\alias{get_data.piece_reg}
\title{Get data function}
\usage{
get_data(model, ...)

\method{get_data}{lm}(model, start = 1, frequency = 1, ...)

\method{get_data}{dynlm}(model, ...)

\method{get_data}{tvlm}(model, end = numeric(), frequency = 1, ...)

\method{get_data}{bp_lm}(model, ...)

\method{get_data}{piece_reg}(model, ...)
}
\arguments{
\item{model}{the model}

\item{...}{other unused parameters.}

\item{start}{the start of the data.}

\item{frequency}{the frequency of the data.}

\item{end}{the end of the data.}
}
\description{
Retrieves the data used in the model
}
