% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_formula}
\alias{get_formula}
\alias{get_formula.default}
\title{Extract Formula From model}
\usage{
get_formula(x)

\method{get_formula}{default}(x)
}
\arguments{
\item{x}{the model.}
}
\description{
Extract Formula From model
}
