% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmse.R
\name{get_rmse}
\alias{get_rmse}
\alias{get_lm_coef}
\alias{get_tvlm_coef}
\alias{get_tvlm_bw}
\alias{get_rmse_bw_small}
\alias{get_coeff_plot}
\title{Get elements of rmse_prev}
\usage{
get_lm_coef(model)

get_tvlm_coef(model, date, variable)

get_tvlm_bw(model)

get_rmse_bw_small(model, ...)

get_coeff_plot(model, date, variable, titre = NULL)
}
\arguments{
\item{model}{result of \link{rmse_prev} fonction}

\item{date}{the date on which we want the tvlm coefficients}

\item{variable}{\code{integer}. The variable of which we want the coefficients at the given date}
}
\description{
Get elements of rmse_prev
}
\details{
\code{get_lm_coeff} allows to get all coefficients of all linear regression prediction models

\code{get_tvlm_coeff} allows to get all coefficients of the variable at a certain date of all local regression prediction models

\code{get_tvlm_bw} allows to get the bandwidth of all local regression prediction models

\code{get_rmse_bw_small} allows to get rmse of linear, local and piecewise regression prediction models, when the bandwidth of the prediction model is different from 20.

\code{get_coeff_plot} plot \code{get_tvlm_coeff} of a certain variable at a certain date, \code{get_lm_coeff} of the same variable and the bandwidth of all prediction models, thanks to \code{get_tvlm_bw}. It also highlights when the bandwidth is equal to 20.
}
