% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data.R
\docType{data}
\name{hansen_table}
\alias{hansen_table}
\title{Hansen Table}
\format{
An object of class \code{data.frame} with 20 rows and 7 columns.
}
\source{
Hansen, Bruce E. 1990. "Lagrange multiplier tests for parameter instability in non-linear models".
\emph{University of Rochester}. \url{https://users.ssc.wisc.edu/~bhansen/papers/LMTests.pdf}.
}
\usage{
hansen_table
}
\description{
Hansen Table
}
\keyword{datasets}
