% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claiRe.R
\name{lm_fenetre_fixe}
\alias{lm_fenetre_fixe}
\title{Computes regression on fixed window}
\usage{
lm_fenetre_fixe(formula, data, nbw = 1)
}
\arguments{
\item{formula}{a \code{formula} object}

\item{data}{time series data}
}
\description{
Computes regression on fixed window
}
\details{
Return an object of class "lmffixe".
Return all models, from which we can extract the usual coefficients, residuals, and fitted.values. And the divisor chosen by the function (arbitrary the middle one), the period, i.e. the length of each sub models, and the frequency of the data.
}
