% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_coef.R
\name{lm_fixed_coeff}
\alias{lm_fixed_coeff}
\title{Fixed and variable coefficients regressions}
\usage{
lm_fixed_coeff(formula, data, var_fixes, ...)
}
\arguments{
\item{formula}{a \code{formula} object.}

\item{data}{time series data.}

\item{var_fixes}{chosen variables whose coefficients aren't allowed to vary through time}

\item{...}{further arguments passed to \code{\link[tvReg:tvLM]{tvReg::tvLM()}}}
}
\value{
\item{global_model}{the simple \code{lm} model}
\item{linear_reg}{simple \code{lm} model with fixed coefficients}
\item{piecewise_reg}{\code{bp.lms} model with fixed coefficients}
\item{tv_reg}{\code{tvlm} model with fixed coefficients}
}
\description{
Computes different types of regressions with some coefficients fixed and others allowed to vary
}
