% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_coef.R
\name{lm_residual_effect}
\alias{lm_residual_effect}
\title{Residual Effect Regression}
\usage{
lm_residual_effect(x, var = c(-1))
}
\arguments{
\item{x}{\code{lm} object}

\item{var}{the variables on which the residuals are to be regressed. By default use them all and cancel the explained variable}
}
\description{
Checks there is no residual effect in a model
}
