% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_reg.R
\name{piece_reg}
\alias{piece_reg}
\title{Piecewise regression}
\usage{
piece_reg(
  x,
  break_dates = NULL,
  var_fixes = NULL,
  tvlm = FALSE,
  bw = NULL,
  left = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{lm} object. It is the global regression model}

\item{break_dates}{optional, to indicate the breakdates if they are known. By default set to \code{NULL}.}

\item{var_fixes}{fixed variables (not splitted using \code{break_dates}).}

\item{tvlm}{By default set to \code{FALSE}. Indicates which model will be run on each sub data. \code{FALSE} means a \link{lm} will be run.}

\item{bw}{bandwidth of the local regression (when \code{tvlm = TRUE}).}

\item{left}{\code{logical}. By default set to \code{TRUE}, i.e. the breakdate is the end date of each submodel}

\item{...}{other arguments passed to \code{\link[tvReg:tvLM]{tvReg::tvLM()}}.}
}
\value{
Returns an element of class \code{lm}
}
\description{
Computes one global linear regression, on splitted data
}
\details{
Computes possible breakdates if not filled in. Uses function \link{break_data} and run a linear regression on the same splitted data.
}
