% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_lm.R
\name{plot.best_ssm}
\alias{plot.best_ssm}
\title{Plot best_ssm}
\usage{
\method{plot}{best_ssm}(x, choice = c("smoothed", "filtered", "filtering"), ...)
}
\arguments{
\item{x}{a \code{best_ssm} object}

\item{choice}{choose between \code{smoothed}, \code{filtered} or \code{filtering}}
}
\description{
Plot explained variable against fitted values computed with smoothed, filtered or filtering states.
}
