% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claiRe.R
\name{rmse_res}
\alias{rmse_res}
\title{Root mean squarred error}
\usage{
rmse_res(resid)
}
\arguments{
\item{resid}{the residuals vector on which rmse will be calculated}
}
\description{
Root mean squarred error
}
