% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soos_prev.R
\name{soos_prev}
\alias{soos_prev}
\alias{soos_prev.lm}
\alias{soos_prev.tvlm}
\alias{soos_prev.bp.lms}
\title{Out of sample prevision (or simulated out of sample)}
\usage{
soos_prev(model, date = 10, period = 1, ...)

\method{soos_prev}{lm}(model, date = 10, period = 1, data, ...)

\method{soos_prev}{tvlm}(
  model,
  data_est = NULL,
  date = 10,
  period = 1,
  fixed_bw = FALSE,
  bw = NULL,
  end,
  frequency,
  ...
)

\method{soos_prev}{bp.lms}(
  model,
  data_est = NULL,
  date = 10,
  period = 1,
  data,
  fixed_bw = FALSE,
  bw = NULL,
  ...
)
}
\arguments{
\item{model}{an object used to select a method}

\item{date}{choose when we want to start the revision process after the start date. By default set to 10.}

\item{period}{choose by how many values we want to move forward. By default set to 1.}

\item{...}{other arguments}
}
\value{
soos_prev returns an object of class \code{revision}, only for models of class \link{lm} and \link{tvlm}. For an object of class \code{bplm} it returns the same previsions and residuals as below.
An object of class \code{revision} is a list containing the following elements :
\item{model}{all models used to forecast}
\item{debut}{same as date chosen earlier}
\item{intervalle}{same as period chosen earlier}
\item{end_dates}{a vector of all end date of each models}
\item{frequency}{the frequency of the data}
\item{prevision}{the forecast}
\item{residuals}{the errors of the forecast}
}
\description{
Out of sample prevision (or simulated out of sample)
}
