% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_lm.R
\name{ssm_lm_best}
\alias{ssm_lm_best}
\title{Best state space model}
\usage{
ssm_lm_best(model)
}
\arguments{
\item{model}{a \code{lm} or \code{dynlm} model}
}
\value{
Return an object of class \code{best_ssm}, a list containing :
\item{rmse_best_model_smoothed}{gives the rmse of the best model for smoothed states}
\item{rmse_best_model_filtering}{gives the rmse of the best model for filtering states}
\item{best_model_smoothed}{the entire best model for smoothed states}
\item{best_model_filtering}{the entire best model for filtering states}
\item{model}{all 25 models}
\item{rmse}{a \code{list} of 3 : all computed rmse}
}
\description{
Computes 25 state space models, using \link{ssm_lm} with different parameters : \code{fixed_intercept} and \code{fixed_variables} either \code{TRUE} or \code{FALSE}, and \code{var_intercept} and \code{var_variables} taking 0.01, 100 or 0 (only when associated parameters is set to \code{TRUE})
}
\details{
It can give different models for smoothed and filtering states.
}
