% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_lm.R
\name{ssm_lm_best_oos}
\alias{ssm_lm_best_oos}
\title{Best out of sample state space model}
\usage{
ssm_lm_best_oos(model)
}
\arguments{
\item{model}{a \code{lm} or \code{dynlm} model}
}
\value{
\item{rmse_best_model_filtering}{gives the rmse of the best model for out of sample prevision}
\item{best_model_filtering}{the best model}
\item{model}{all 25 models}
\item{rmse_filtering}{all 25 rmse}
}
\description{
Computes 25 state space models, using \link{ssm_lm_oos} with different parameters : \code{fixed_var_intercept} and \code{fixed_var_variables} either \code{TRUE} or \code{FALSE}, and \code{var_intercept} and \code{var_variables} taking 0.01, 100 or 0 (only when associated parameters is set to \code{TRUE}).

Is is the same 25 conbinations of parameters than in \link{ssm_lm_best}
}
\details{
It can give different models for smoothed and filtering states.
}
