% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\alias{get_data.lm}
\alias{get_data.dynlm}
\alias{get_data.tvlm}
\alias{get_data.bp_lm}
\alias{get_data.piece_reg}
\title{Get data function}
\usage{
get_data(model, ...)

\method{get_data}{lm}(model, start, ...)

\method{get_data}{dynlm}(model, ...)

\method{get_data}{tvlm}(model, end, frequency, ...)

\method{get_data}{bp_lm}(model, ...)

\method{get_data}{piece_reg}(model, ...)
}
\arguments{
\item{model}{the model whose data we want}
}
\description{
Retrieves the data used in the model
}
