% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data.R
\docType{data}
\name{hansen_table}
\alias{hansen_table}
\title{Hansen Table}
\format{
An object of class \code{data.frame} with 20 rows and 7 columns.
}
\usage{
hansen_table
}
\description{
Hansen Table
}
\keyword{datasets}
