% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data.R
\docType{data}
\name{manufacturing}
\alias{manufacturing}
\title{Business Surveys}
\format{
An object of class \code{mts} (inherits from \code{ts}, \code{matrix}) with 300 rows and 208 columns.
}
\source{
INSEE, Banque de France
}
\usage{
manufacturing
}
\description{
Business Surveys
}
\keyword{datasets}
