% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soos_prev.R
\name{oos_prev}
\alias{oos_prev}
\alias{oos_prev.lm}
\alias{oos_prev.tvlm}
\alias{oos_prev.bp_lm}
\alias{oos_prev.piece_reg}
\title{Out of sample forecast (or simulated out of sample)}
\usage{
oos_prev(model, date = 28, period = 1, ...)

\method{oos_prev}{lm}(model, date = 28, period = 1, data, ...)

\method{oos_prev}{tvlm}(
  model,
  date = 28,
  period = 1,
  data_est = NULL,
  fixed_bw = FALSE,
  bw = NULL,
  end,
  frequency,
  ...
)

\method{oos_prev}{bp_lm}(
  model,
  date = 28,
  period = 1,
  data_est = NULL,
  data,
  fixed_bw = FALSE,
  bw = NULL,
  ...
)

\method{oos_prev}{piece_reg}(model, date = 28, period = 1, ...)
}
\arguments{
\item{model}{an object used to select a method}

\item{date}{choose when we want to start the revision process after the start date. By default set to 28 periods.}

\item{period}{choose by how many values we want to move forward. By default set to 1.}

\item{...}{other arguments}
}
\value{
oos_prev returns an object of class \code{revision}, only for models of class \link{lm} and \link{tvlm}. For an object of class \code{bplm} it returns the same forecasts and residuals as below.
An object of class \code{revision} is a list containing the following elements:
\item{model}{all models used to forecast}
\item{debut}{same as date chosen earlier}
\item{intervalle}{same as period chosen earlier}
\item{end_dates}{a vector of all end date of each models}
\item{frequency}{the frequency of the data}
\item{forecast}{the forecast}
\item{residuals}{the errors of the forecast}
}
\description{
Out of sample forecast (or simulated out of sample)
}
