% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_reg.R
\name{piece_reg}
\alias{piece_reg}
\title{Piecewise regression}
\usage{
piece_reg(
  x,
  break_dates = NULL,
  var_fixes = NULL,
  tvlm = FALSE,
  bw = NULL,
  left = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{lm} object, of which we will separate the data}

\item{break_dates}{optional, to indicate the breakdates if they are known. By default set to \code{NULL}}
}
\value{
Returns an element of class \code{lm}
}
\description{
Computes one global linear regression, on splitted data
}
\details{
Computes possible breakdates if not filled in. Uses function \link{break_data} and run a linear regression on the same splitted data.
}
