% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_lm.R
\name{ssm_lm}
\alias{ssm_lm}
\title{State space model}
\usage{
ssm_lm(
  x,
  trend = FALSE,
  var_intercept = 0,
  var_slope = 0,
  var_variables = 0,
  fixed_intercept = TRUE,
  fixed_trend = TRUE,
  fixed_variables = TRUE,
  ...,
  remove_last_dummies = FALSE
)
}
\arguments{
\item{trend}{trend}

\item{var_intercept, var_slope}{variance of the intercept and the slope (used if \code{trend = TRUE}).}

\item{var_variables}{??}

\item{fixed_intercept}{??}

\item{fixed_variables}{??}

\item{model}{a \code{lm} or \code{dynlm} model}
}
\value{
Returns a \code{list} containig :
\item{smoothed_states}{\eqn{E[a_t|y_0,\dots,y_n]}}
\item{smoothed_var}{\eqn{V[a_t|y_0,\dots,y_n]}}
\item{filtered_states}{\eqn{E[a_t|y_0,\dots,y_t]}}
\item{filtered_var}{\eqn{E[a_t|y_0,\dots,y_t]}}
\item{filtering_states}{\eqn{E[a_t|y_0,\dots,y_{t-1}]}}
\item{filtering_var}{\eqn{V[a_t|y_0,\dots,y_{t-1}]}}
\item{data}{data used in the original model}
}
\description{
Computes state space model with one equation. Starting with a simple \code{lm} model, build the all state space model and run it.

Use \link{rjd3stst} and \link{rjd3toolkit} packages.
}
